reorganize the React folder
