+++ react/calendar-event-editor.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import CRMEvent from './interfaces/CRMEvent';
-import ExistingEvent from './components/ExistingEvent';
+import ExistingEvent from './components/Events/ExistingEvent';
 declare global {
     interface Moment {
       _isAMomentObject: boolean,
