+++ react/components/Events/EventPropertiesEditor.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import CRMEvent from '../interfaces/CRMEvent';
-import Calendar from '../interfaces/Calendar';
-import EventType from '../interfaces/EventType';
+import CRMEvent from '../../interfaces/CRMEvent';
+import Calendar from '../../interfaces/Calendar';
+import EventType from '../../interfaces/EventType';
 import { Modal, FormControl } from 'react-bootstrap';
 import Select from 'react-select';
 import DatePicker from 'react-datepicker';
