+++ react/components/Events/ExistingEvent.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
-import CRMEvent from '../interfaces/CRMEvent';
-import Calendar from '../interfaces/Calendar';
-import EventType from '../interfaces/EventType';
+import CRMEvent from '../../interfaces/CRMEvent';
+import Calendar from '../../interfaces/Calendar';
+import EventType from '../../interfaces/EventType';
 import { Modal, FormControl } from 'react-bootstrap';
-import CRMRoot from '../window-context-service.jsx';
+import CRMRoot from '../../window-context-service.jsx';
 import EventPropertiesViewer from './EventPropertiesViewer';
 import EventPropertiesEditor from './EventPropertiesEditor';
